crap

+++ db/migrate/20180711195455_create_points.rb
@@ -1,8 +1,8 @@
 class CreatePoints < ActiveRecord::Migration[5.2]
   def change
     create_table :points do |t|
-      t.float :x
-      t.float :y
+      t.integer :x
+      t.integer :y
       t.references :track, index: true
 
       t.timestamps
